+++ src/codec/gzip/decoder.rs
@@ -81,18 +81,14 @@ impl GzipDecoder {
                 State::Decoding => {
                     let prior = output.written().len();
 
-                    let done = inner(self, input, output).or_else(|e| match e {
-                        // we need to make an exception, if output flushed
-                        // from the decoder, to update the crc
-                        e if e.kind() == std::io::ErrorKind::Other
-                            && output.written().len() > 0 =>
-                        {
-                            Ok(false)
-                        }
-                        _ => Err(e),
-                    })?;
+                    let res = inner(self, input, output);
 
-                    self.crc.update(&output.written()[prior..]);
+                    if (output.written().len() > prior) {
+                        // update CRC even if there was an error
+                        self.crc.update(&output.written()[prior..]);
+                    }
+
+                    let done = res?;
 
                     if done {
                         self.state = State::Footer(vec![0; 8].into())